+++ tests/test_utils.py
@@ -262,6 +262,7 @@ def test_temporary_docker_directory_uses_hard_link():
             static=[],
             install=[],
             spatialite=False,
+            slim_base_image=False,
             version_note=None,
             secret="secret",
         ) as temp_docker:
@@ -292,6 +293,7 @@ def test_temporary_docker_directory_uses_copy_if_hard_link_fails(mock_link):
             static=[],
             install=[],
             spatialite=False,
+            slim_base_image=False,
             version_note=None,
             secret=None,
         ) as temp_docker:
@@ -318,6 +320,7 @@ def test_temporary_docker_directory_quotes_args():
             static=[],
             install=[],
             spatialite=False,
+            slim_base_image=False,
             version_note="$PWD",
             secret="secret",
         ) as temp_docker:
@@ -327,6 +330,54 @@ def test_temporary_docker_directory_quotes_args():
             assert "'$PWD'" in df_contents
             assert "'--$HOME'" in df_contents
             assert "ENV DATASETTE_SECRET 'secret'" in df_contents
+
+
+def test_slim_base_image():
+    with tempfile.TemporaryDirectory() as td:
+        os.chdir(td)
+        with utils.temporary_docker_directory(
+            files=["hello"],
+            name="t",
+            metadata=None,
+            extra_options="--$HOME",
+            branch=None,
+            template_dir=None,
+            plugins_dir=None,
+            static=[],
+            install=[],
+            spatialite=False,
+            slim_base_image=True,
+            version_note="$PWD",
+            secret="secret",
+        ) as temp_docker:
+            df = os.path.join(temp_docker, "Dockerfile")
+            with open(df) as fp:
+                df_contents = fp.read()
+            assert "FROM python:3.8-slim-bullseye" in df_contents
+
+
+def test_fat_base_image():
+    with tempfile.TemporaryDirectory() as td:
+        os.chdir(td)
+        with utils.temporary_docker_directory(
+            files=["hello"],
+            name="t",
+            metadata=None,
+            extra_options="--$HOME",
+            branch=None,
+            template_dir=None,
+            plugins_dir=None,
+            static=[],
+            install=[],
+            spatialite=False,
+            slim_base_image=False,
+            version_note="$PWD",
+            secret="secret",
+        ) as temp_docker:
+            df = os.path.join(temp_docker, "Dockerfile")
+            with open(df) as fp:
+                df_contents = fp.read()
+            assert "FROM python:3.8\n" in df_contents
 
 
 def test_compound_keys_after_sql():