+++ docs/changelog.rst
@@ -21,6 +21,7 @@ Changelog
 0.60a0 (2021-12-17)
 -------------------
 
+- New option ``--slim-base-images`` for ``package`` command that uses the python slim variant.
 - New plugin hook: :ref:`plugin_hook_filters_from_request`, which runs on the table page and can be used to support new custom query string parameters that modify the SQL query. (:issue:`473`)
 - The number of unique values in a facet is now always displayed. Previously it was only displayed if the user specified ``?_facet_size=max``. (:issue:`1556`)
 - Fixed bug where ``?_facet_array=tags&_facet=tags`` would only display one of the two selected facets. (:issue:`625`)